feat(): moved the functionality to `system_service`, leaving only the orchestration function in `tool_kill_service`. Organized the comments. Added a safeguard that, in the event of a system freeze, will trigger a forced reinstallation after 10 seconds to avoid an interruption lasting about 4 minutes.
