+++ clients/openframe-client/src/services/tool_kill_service.rs
@@ -1,15 +1,13 @@
-use anyhow::{Context, Result};
+use anyhow::Result;
 use tracing::{info, warn, error};
 use sysinfo::{System, Signal, Pid};
 use tokio::time::{sleep, Duration};
-use tokio::process::Command;
 use crate::models::{InstalledTool, Installation};
+use crate::platform::system_service;
 use crate::config::service_stop::{
     FORCE_KILL_TIMEOUT_SECS, GRACEFUL_SHUTDOWN_TIMEOUT_SECS, MAX_KILL_RETRIES,
     PROCESS_CHECK_INTERVAL_MS,
 };
-#[cfg(target_os = "windows")]
-use crate::config::service_stop::{SERVICE_FORCE_KILL_MAX_ATTEMPTS, SERVICE_STOP_MAX_ATTEMPTS};
 
 /// Service responsible for stopping/killing tool processes
 #[derive(Clone)]
@@ -224,11 +222,7 @@ impl ToolKillService {
             Installation::Service { service_name, executable_path } => {
                 info!(service_name = %service_name,
                       "Stopping Service type tool via system service manager");
-                // Non-fatal: even if the service manager can't stop it, fall through
-                // to the path-based process kill below so a stuck/pending service
-                // (e.g. Mesh Agent returning error 1061) doesn't leave the binary
-                // locked and dead-lock the reinstall/update.
-                if let Err(e) = self.stop_service(service_name).await {
+                if let Err(e) = system_service::stop_service(service_name).await {
                     warn!("Failed to stop service {} (continuing with process kill by path): {:#}",
                           service_name, e);
                 }
@@ -268,280 +262,4 @@ impl ToolKillService {
             format!("/{}/{}", tool_id, asset_id).to_lowercase()
         }
     }
-
-    pub async fn stop_service(&self, service_name: &str) -> Result<()> {
-        info!("Stopping service: {}", service_name);
-
-        #[cfg(target_os = "windows")]
-        {
-            self.stop_service_windows(service_name).await
-        }
-
-        #[cfg(target_os = "macos")]
-        {
-            self.stop_service_macos(service_name).await
-        }
-
-        #[cfg(target_os = "linux")]
-        {
-            self.stop_service_linux(service_name).await
-        }
-    }
-
-    #[cfg(target_os = "windows")]
-    async fn stop_service_windows(&self, service_name: &str) -> Result<()> {
-        use windows_service::service::ServiceAccess;
-        use winapi::shared::winerror::{
-            ERROR_SERVICE_CANNOT_ACCEPT_CTRL, ERROR_SERVICE_DOES_NOT_EXIST, ERROR_SERVICE_NOT_ACTIVE,
-        };
-
-        info!("Stopping Windows service via SCM: {}", service_name);
-
-        let service = match open_service_windows(
-            service_name,
-            ServiceAccess::QUERY_STATUS | ServiceAccess::STOP,
-        ) {
-            Ok(service) => service,
-            // The specified service does not exist — nothing to stop.
-            Err(windows_service::Error::Winapi(e))
-                if e.raw_os_error() == Some(ERROR_SERVICE_DOES_NOT_EXIST as i32) =>
-            {
-                warn!("Service {} does not exist (error {})", service_name, ERROR_SERVICE_DOES_NOT_EXIST);
-                return Ok(());
-            }
-            Err(e) => {
-                warn!("Failed to open service {} to stop it: {}; force-killing", service_name, e);
-                return self.force_stop_service_windows(service_name).await;
-            }
-        };
-
-        match service.stop() {
-            Ok(_) => {
-                info!("Service {} stop initiated", service_name);
-                // stop() only *initiates* the stop; the SCM may keep reporting the
-                // service as running for a while. If it never reaches STOPPED, fall
-                // back to force-killing so we don't leave the binary locked.
-                if self.wait_for_service_stop_windows(service_name).await? {
-                    return Ok(());
-                }
-                warn!("Service {} did not reach STOPPED after stop request; force-killing", service_name);
-                self.force_stop_service_windows(service_name).await
-            }
-            // The service has not been started.
-            Err(windows_service::Error::Winapi(e))
-                if e.raw_os_error() == Some(ERROR_SERVICE_NOT_ACTIVE as i32) =>
-            {
-                info!("Service {} is not running (error {})", service_name, ERROR_SERVICE_NOT_ACTIVE);
-                Ok(())
-            }
-            // ERROR_SERVICE_CANNOT_ACCEPT_CTRL: the service is in a pending/
-            // transitional state and won't accept a stop right now. This is common
-            // with the Mesh Agent on Windows Server and previously caused the
-            // reinstall to abort, leaving agent.exe locked (os error 32) in a retry
-            // loop. Force-kill so the reinstall/update can proceed.
-            Err(windows_service::Error::Winapi(e))
-                if e.raw_os_error() == Some(ERROR_SERVICE_CANNOT_ACCEPT_CTRL as i32) =>
-            {
-                warn!("Service {} cannot accept stop control (error {}); force-killing service process",
-                      service_name, ERROR_SERVICE_CANNOT_ACCEPT_CTRL);
-                self.force_stop_service_windows(service_name).await
-            }
-            // Any other failure: force-kill rather than dead-lock the reinstall.
-            Err(e) => {
-                error!("Failed to stop service {} via SCM: {}; force-killing service process", service_name, e);
-                self.force_stop_service_windows(service_name).await
-            }
-        }
-    }
-
-    /// Force-stop a Windows service by killing its host process tree.
-    ///
-    /// Used when the SCM can't gracefully stop the service (e.g. error 1061, the
-    /// service is in a pending state). Reads the host PID from the SCM
-    /// (`query_status().process_id`, only populated while RUNNING) and terminates
-    /// it with `taskkill /F /T`, retrying to ride out pending transitions and to
-    /// defeat SCM auto-restart, until the service is no longer running.
-    #[cfg(target_os = "windows")]
-    async fn force_stop_service_windows(&self, service_name: &str) -> Result<()> {
-        for attempt in 1..=SERVICE_FORCE_KILL_MAX_ATTEMPTS {
-            let status = query_service_status_windows(service_name);
-            if service_stopped_or_missing(&status) {
-                info!("Service {} is no longer running (force-stop attempt {})", service_name, attempt);
-                return Ok(());
-            }
-
-            match status.as_ref().ok().and_then(|s| s.process_id) {
-                Some(pid) => {
-                    info!("Force-killing service {} process tree (pid {}, attempt {}/{})",
-                          service_name, pid, attempt, SERVICE_FORCE_KILL_MAX_ATTEMPTS);
-
-                    let output = Command::new("taskkill")
-                        .args(["/F", "/T", "/PID", &pid.to_string()])
-                        .output()
-                        .await
-                        .with_context(|| format!("Failed to execute taskkill for service {} (pid {})", service_name, pid))?;
-
-                    if !output.status.success() {
-                        let kill_stdout = String::from_utf8_lossy(&output.stdout);
-                        let kill_stderr = String::from_utf8_lossy(&output.stderr);
-                        // taskkill returns non-zero if the process already exited
-                        // between query and kill; that's fine, the next check confirms.
-                        warn!("taskkill for service {} (pid {}) reported: stdout={} stderr={}",
-                              service_name, pid, kill_stdout.trim(), kill_stderr.trim());
-                    }
-                }
-                None => {
-                    // The SCM only reports a PID while RUNNING, so no PID means the
-                    // service is mid-transition (start/stop pending). Wait for it to
-                    // settle; the path-based kill in `stop_for_installation` is the
-                    // backstop if it never yields a killable PID.
-                    let state = status.as_ref().map(|s| format!("{:?}", s.current_state))
-                        .unwrap_or_else(|_| "unqueryable".to_string());
-                    info!("Service {} has no reportable PID (state {}, attempt {}/{}); waiting",
-                          service_name, state, attempt, SERVICE_FORCE_KILL_MAX_ATTEMPTS);
-                }
-            }
-
-            sleep(Duration::from_millis(PROCESS_CHECK_INTERVAL_MS)).await;
-        }
-
-        // Final verification.
-        let status = query_service_status_windows(service_name);
-        if service_stopped_or_missing(&status) {
-            info!("Service {} force-stopped successfully", service_name);
-            Ok(())
-        } else {
-            let state = status.as_ref().map(|s| format!("{:?}", s.current_state))
-                .unwrap_or_else(|_| "unqueryable".to_string());
-            error!("Service {} still not stopped (state {}) after {} force-kill attempts",
-                   service_name, state, SERVICE_FORCE_KILL_MAX_ATTEMPTS);
-            Err(anyhow::anyhow!(
-                "Failed to force-stop service {} (state {} after {} attempts)",
-                service_name, state, SERVICE_FORCE_KILL_MAX_ATTEMPTS
-            ))
-        }
-    }
-
-    /// Poll the SCM until the service reports STOPPED (or no longer exists).
-    ///
-    /// Returns `Ok(true)` if it reached STOPPED/removed, `Ok(false)` on timeout.
-    #[cfg(target_os = "windows")]
-    async fn wait_for_service_stop_windows(&self, service_name: &str) -> Result<bool> {
-        for attempt in 1..=SERVICE_STOP_MAX_ATTEMPTS {
-            sleep(Duration::from_millis(PROCESS_CHECK_INTERVAL_MS)).await;
-
-            if service_stopped_or_missing(&query_service_status_windows(service_name)) {
-                info!("Service {} confirmed stopped after {} attempts", service_name, attempt);
-                return Ok(true);
-            }
-        }
-
-        warn!("Service {} did not confirm stopped after {} attempts", service_name, SERVICE_STOP_MAX_ATTEMPTS);
-        Ok(false)
-    }
-
-    #[cfg(target_os = "macos")]
-    async fn stop_service_macos(&self, service_name: &str) -> Result<()> {
-        let plist_path = format!("/Library/LaunchDaemons/{}.plist", service_name);
-        info!("Stopping macOS service via sudo launchctl unload: {}", plist_path);
-
-        // Check if plist exists
-        if !std::path::Path::new(&plist_path).exists() {
-            warn!("Plist not found at {}, service may not be installed", plist_path);
-            return Ok(());
-        }
-
-        let output = Command::new("sudo")
-            .args(["launchctl", "unload", &plist_path])
-            .output()
-            .await
-            .with_context(|| format!("Failed to execute sudo launchctl unload for: {}", plist_path))?;
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-
-        if output.status.success() {
-            info!("Service unloaded successfully: {}", plist_path);
-            Ok(())
-        } else if stderr.contains("Could not find specified service") {
-            info!("Service not loaded (already stopped): {}", plist_path);
-            Ok(())
-        } else if stderr.contains("No such file or directory") {
-            warn!("Plist not found: {}", plist_path);
-            Ok(())
-        } else {
-            error!("Failed to unload service {}: {}", plist_path, stderr);
-            Err(anyhow::anyhow!(
-                "Failed to unload service {}: {}",
-                plist_path,
-                stderr
-            ))
-        }
-    }
-
-    #[cfg(target_os = "linux")]
-    async fn stop_service_linux(&self, service_name: &str) -> Result<()> {
-        info!("Stopping Linux service via systemctl stop: {}", service_name);
-
-        let output = Command::new("systemctl")
-            .args(["stop", service_name])
-            .output()
-            .await
-            .with_context(|| format!("Failed to execute systemctl stop for service: {}", service_name))?;
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-
-        if output.status.success() {
-            info!("Service {} stopped successfully", service_name);
-            Ok(())
-        } else if stderr.contains("not loaded") || stderr.contains("not found") {
-            warn!("Service {} not found or not loaded", service_name);
-            Ok(())
-        } else {
-            error!("Failed to stop service {}: {}", service_name, stderr);
-            Err(anyhow::anyhow!(
-                "Failed to stop service {}: {}",
-                service_name,
-                stderr
-            ))
-        }
-    }
 }
-
-/// Open a Windows service handle with the requested access via the SCM.
-#[cfg(target_os = "windows")]
-fn open_service_windows(
-    service_name: &str,
-    access: windows_service::service::ServiceAccess,
-) -> windows_service::Result<windows_service::service::Service> {
-    use windows_service::service_manager::{ServiceManager, ServiceManagerAccess};
-    let manager = ServiceManager::local_computer(None::<&str>, ServiceManagerAccess::CONNECT)?;
-    manager.open_service(service_name, access)
-}
-
-/// Query a Windows service's current status via the SCM.
-#[cfg(target_os = "windows")]
-fn query_service_status_windows(
-    service_name: &str,
-) -> windows_service::Result<windows_service::service::ServiceStatus> {
-    use windows_service::service::ServiceAccess;
-    let service = open_service_windows(service_name, ServiceAccess::QUERY_STATUS)?;
-    service.query_status()
-}
-
-/// True when the service is STOPPED or no longer exists — i.e. there's nothing
-/// left to kill. A transient query error is treated as "not yet stopped".
-#[cfg(target_os = "windows")]
-fn service_stopped_or_missing(
-    status: &windows_service::Result<windows_service::service::ServiceStatus>,
-) -> bool {
-    use winapi::shared::winerror::ERROR_SERVICE_DOES_NOT_EXIST;
-    use windows_service::service::ServiceState;
-    match status {
-        Ok(s) => s.current_state == ServiceState::Stopped,
-        Err(windows_service::Error::Winapi(e)) => {
-            e.raw_os_error() == Some(ERROR_SERVICE_DOES_NOT_EXIST as i32)
-        }
-        Err(_) => false,
-    }
-}
-
